requirements.txt
flask==3.0.3
flask-restx==1.3.0
flask-cors==4.0.1
sqlalchemy==2.0.32
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
